backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
beanie==1.26.0
motor==3.6.0
pydantic-settings==2.6.1

PyJWT==2.9.0
requests==2.32.3

chromadb==1.5.0
posthog
sentence-transformers==3.1.1

passlib[bcrypt]==1.7.4

httpx==0.27.2

motor==3.6.0
sse-starlette==2.1.3
python-dateutil==2.9.0.post0



# SCIM helpers
scim2-filter-parser==0.7.0

pydevd-pycharm~=253.30387.90
